requirements.txt
fastmcp
python-dotenv
fastapi
httpx
pytest
pre-commit
ruff
airportsdata
geopy
